build(cxx23): Phase 2 — match C warning policy on C++ side; fix enum-ternary
- Apply ${_gcc_only_flags} (-Wno-format-truncation, -Wno-unused-result,
  -Wno-stringop-truncation, -Wno-alloc-size-larger-than) to
  CBM_FIRSTPARTY_CXX_FLAGS, mirroring the C flags. These guard the codebase's
  intentional safe-snprintf-truncation and a couple of deliberately-ignored
  fread/freopen returns; the C build already suppressed them. Clears ~40 C++
  warnings that the clean build had surfaced.
- Cast the anonymous-enum arm of five `cond ? (size_t)x : ENUM` ternaries to
  size_t (cypher x2, mem, pipeline_incremental, semantic), clearing the
  -Wextra "enumerated and non-enumerated type in conditional" warnings.

Remaining after this: 9 -Wunused-function warnings in ts_lsp/cs_lsp, left
visible on purpose — they flag resolver helpers that are never called
(possible incomplete wiring vs dead code) and want a maintainer's triage,
not a silent delete or mask.

Test parity: 3629/1 with ASan+UBSan (the known RSS check only).

+++ src/cypher/cypher.cpp
@@ -2183,7 +2183,7 @@ static void rb_init(result_builder_t *rb) {
 }
 
 static void rb_set_columns(result_builder_t *rb, const char **cols, int count) {
-    rb->columns = (__typeof__(rb->columns))malloc((count > 0 ? (size_t)count : SKIP_ONE) * sizeof(const char *));
+    rb->columns = (__typeof__(rb->columns))malloc((count > 0 ? (size_t)count : (size_t)SKIP_ONE) * sizeof(const char *));
     for (int i = 0; i < count; i++) {
         rb->columns[i] = heap_strdup(cols[i]);
     }
@@ -2195,7 +2195,7 @@ static void rb_add_row(result_builder_t *rb, const char **values) {
         rb->row_cap *= PAIR_LEN;
         rb->rows = (__typeof__(rb->rows))safe_realloc(rb->rows, rb->row_cap * sizeof(const char **));
     }
-    const char **row = (const char **)malloc((rb->col_count > 0 ? (size_t)rb->col_count : SKIP_ONE) * sizeof(const char *));
+    const char **row = (const char **)malloc((rb->col_count > 0 ? (size_t)rb->col_count : (size_t)SKIP_ONE) * sizeof(const char *));
     for (int i = 0; i < rb->col_count; i++) {
         row[i] = values[i] ? heap_strdup(values[i]) : heap_strdup("");
     }
